NoCodeTune/FineTuneFlow/requirements.txt
uvicorn[standard]
fastapi
transformers
torch
accelerate
sentencepiece
python-dotenv=
peft
pandas
python-docx
PyPDF2
markdown
bs4
pylatexenc
colorama
python-multipart
openpyxl
datasets
swifter
SQLAlchemy
psycopg2
psutil
Jinja2
pydantic-argparse
langchain_community